Move LockFlags into client/application side
