+++ radix-engine-interface/Cargo.toml
@@ -12,6 +12,7 @@ bech32 = { version = "0.9.0", default-features = false }
 sbor = { path = "../sbor", default-features = false }
 radix-engine-derive = { path = "../radix-engine-derive", default-features = false }
 scrypto-abi = { path = "../scrypto-abi", default-features = false }
+bitflags = { version = "1.3" }
 utils = { path = "../utils", default-features = false }
 transaction-data = { path = "../transaction-data", default-features = false }
 paste = { version = "1.0.7"}
