+++ radix-engine/src/blueprints/logger/package.rs
@@ -1,5 +1,5 @@
 use crate::errors::{InterpreterError, RuntimeError};
-use crate::kernel::kernel_api::LockFlags;
+use radix_engine_interface::api::substate_api::LockFlags;
 use crate::kernel::kernel_api::{KernelNodeApi, KernelSubstateApi};
 use crate::system::kernel_modules::costing::FIXED_LOW_FEE;
 use crate::types::*;
