+++ radix-engine/src/kernel/call_frame.rs
@@ -1,6 +1,5 @@
 use crate::errors::{CallFrameError, KernelError, RuntimeError};
 use crate::kernel::actor::ResolvedActor;
-use crate::kernel::kernel_api::LockFlags;
 use crate::system::node::{RENodeInit, RENodeModuleInit};
 use crate::system::node_properties::SubstateProperties;
 use crate::system::node_substates::{SubstateRef, SubstateRefMut};
@@ -8,6 +7,7 @@ use crate::types::*;
 use radix_engine_interface::api::types::{
     Address, LockHandle, NonFungibleStoreOffset, RENodeId, SubstateId, SubstateOffset,
 };
+use radix_engine_interface::api::substate_api::LockFlags;
 
 use super::heap::{Heap, HeapRENode};
 use super::kernel_api::LockInfo;
