+++ radix-engine/src/kernel/module_mixer.rs
@@ -2,7 +2,6 @@ use super::actor::ResolvedActor;
 use super::kernel_api::KernelModuleApi;
 use crate::errors::*;
 use crate::kernel::call_frame::CallFrameUpdate;
-use crate::kernel::kernel_api::LockFlags;
 use crate::kernel::module::KernelModule;
 use crate::system::kernel_modules::auth::AuthModule;
 use crate::system::kernel_modules::costing::CostingModule;
@@ -20,6 +19,7 @@ use crate::system::node::RENodeInit;
 use crate::system::node::RENodeModuleInit;
 use crate::types::api::unsafe_api::ClientCostingReason;
 use bitflags::bitflags;
+use radix_engine_interface::api::substate_api::LockFlags;
 use radix_engine_interface::api::types::InvocationIdentifier;
 use radix_engine_interface::api::types::LockHandle;
 use radix_engine_interface::api::types::NodeModuleId;
