+++ radix-engine/src/system/kernel_modules/kernel_debug/module.rs
@@ -1,11 +1,11 @@
 use crate::kernel::actor::ResolvedActor;
 use crate::kernel::call_frame::CallFrameUpdate;
-use crate::kernel::kernel_api::LockFlags;
 use crate::{
     errors::RuntimeError,
     kernel::{kernel_api::KernelModuleApi, module::KernelModule},
     system::node::{RENodeInit, RENodeModuleInit},
 };
+use radix_engine_interface::api::substate_api::LockFlags;
 use radix_engine_interface::api::types::{
     InvocationIdentifier, LockHandle, NodeModuleId, RENodeId, RENodeType, SubstateOffset,
 };
